Separated the samples and tests per Justin's comments.
