+++ samples/system-test/vision_async_batch_annotate_images_beta.test.js
@@ -24,7 +24,7 @@ const uuid = require('uuid');
 const exec = async cmd => (await execa.shell(cmd)).stdout;
 const storage = new Storage();
 const bucketName = `nodejs-docs-samples-test-${uuid.v4()}`;
-const cmd = `node detect.v1p4beta1.js`;
+const cmd = `node vision_async_batch_annotate_images_beta.js`;
 
 const files = [`pdf-ocr.pdf`, `landmark.jpg`].map(name => {
   return {
@@ -33,7 +33,7 @@ const files = [`pdf-ocr.pdf`, `landmark.jpg`].map(name => {
   };
 });
 
-describe(`detect v1 p4 beta1`, () => {
+describe(`Detects batch annotation of images on Google Cloud Storage asynchronously`, () => {
   before(async () => {
     const [bucket] = await storage.createBucket(bucketName);
     await Promise.all(files.map(file => bucket.upload(file.localPath)));
@@ -45,27 +45,9 @@ describe(`detect v1 p4 beta1`, () => {
     await bucket.delete();
   });
 
-  it(`should annotate the local pdf-ocr.pdf sample`, async () => {
-    const output = await exec(
-      `${cmd} detectBatchAnnotateFiles ${files[0].localPath}`
-    );
-    assert.match(output, /Word text: Boring/);
-    assert.match(output, /Symbol: p/);
-  });
-
-  it(`should annotate the remote pdf-ocr.pdf in GCS bucket`, async () => {
-    const output = await exec(
-      `${cmd} detectBatchAnnotateFilesGCS gs://${bucketName}/${files[0].name}`
-    );
-    assert.match(output, /Word text: Boring/);
-    assert.match(output, /Symbol: p/);
-  });
-
   it(`should annotate the remote landmark.jpg sample`, async () => {
     const output = await exec(
-      `${cmd} detectBatchAnnotateImageUri gs://${bucketName}/${
-        files[1].name
-      } gs://${bucketName}/out/`
+      `${cmd} gs://${bucketName}/${files[1].name} gs://${bucketName}/out/`
     );
     assert.match(output, /Json saved to: gs:\/\//);
   });
