+++ samples/system-test/vision_batch_annotate_files_gcs_beta.test.js
@@ -1,0 +1,53 @@
+/**
+ * Copyright 2019, Google, LLC.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+'use strict';
+
+const path = require('path');
+const {Storage} = require('@google-cloud/storage');
+const execa = require('execa');
+const {assert} = require('chai');
+const uuid = require('uuid');
+
+const exec = async cmd => (await execa.shell(cmd)).stdout;
+const storage = new Storage();
+const bucketName = `nodejs-docs-samples-test-${uuid.v4()}`;
+const cmd = `node vision_batch_annotate_files_gcs_beta.js`;
+
+const files = [`pdf-ocr.pdf`, `landmark.jpg`].map(name => {
+  return {
+    name,
+    localPath: path.resolve(path.join(__dirname, `../resources/${name}`)),
+  };
+});
+
+describe(`Detects annotations in a GCS of file`, () => {
+  before(async () => {
+    const [bucket] = await storage.createBucket(bucketName);
+    await Promise.all(files.map(file => bucket.upload(file.localPath)));
+  });
+
+  after(async () => {
+    const bucket = storage.bucket(bucketName);
+    await bucket.deleteFiles({force: true});
+    await bucket.delete();
+  });
+
+  it(`should annotate the remote pdf-ocr.pdf in GCS bucket`, async () => {
+    const output = await exec(`${cmd} gs://${bucketName}/${files[0].name}`);
+    assert.match(output, /Word text: Boring/);
+    assert.match(output, /Symbol: p/);
+  });
+});
